# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DartShop API Documentation",
        "description": "API documentation for DartShop Laravel application",
        "contact": {
            "name": "DartShop Support",
            "email": "admin@dartshop.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/admin/reviews": {
            "get": {
                "tags": [
                    "Admin/Reviews"
                ],
                "summary": "Get reviews list (admin)",
                "description": "Retrieve all reviews with admin filters and pagination",
                "operationId": "9a0aa27eebbefe2594729b4f6e750965",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search in title, content, user name, product name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Filter by rating (1-5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 5,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "approved",
                        "in": "query",
                        "description": "Filter by approval status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "description": "Filter by featured status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "sort_field",
                        "in": "query",
                        "description": "Sort field (id, rating, created_at, product)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc, desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Review"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Admin access required"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "to": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 60
                    },
                    "cache_hit": {
                        "type": "boolean",
                        "example": false
                    },
                    "cache_key": {
                        "type": "string",
                        "example": "products_list_abc123..."
                    },
                    "filters_available": {
                        "properties": {
                            "categories": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            },
                            "brands": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Brand"
                                }
                            },
                            "price_range": {
                                "properties": {
                                    "min": {
                                        "type": "number",
                                        "example": 10
                                    },
                                    "max": {
                                        "type": "number",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "title": "Pagination Links",
                "description": "Pagination links",
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=5"
                    },
                    "prev": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "example": "http://example.com/api/products?page=2"
                    }
                },
                "type": "object"
            },
...
            "Review": {
                "title": "Review",
                "description": "Product review model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Great product!"
                    },
                    "content": {
                        "type": "string",
                        "example": "This is an excellent product..."
                    },
                    "rating": {
                        "type": "integer",
                        "example": 5
                    },
                    "is_approved": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "product": {
                        "$ref": "#/components/schemas/Product"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
            "Brand": {
                "title": "Brand",
                "description": "Product brand model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Harrows"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Product category model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Dart Flights"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Harrows Supergrip Ultra"
                    },
                    "description": {
                        "type": "string",
                        "example": "Professional dart flights"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": false
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "brand": {
                        "$ref": "#/components/schemas/Brand"
                    },
                    "average_rating": {
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "reviews_count": {
                        "type": "integer",
                        "example": 10
                    },
                    "promotion_price": {
                        "type": "number",
                        "format": "float",
                        "example": 24.99,
                        "nullable": true
                    },
                    "savings": {
                        "type": "number",
                        "format": "float",
                        "example": 5
                    },
                    "promotion": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Promotion"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Promotion": {
                "title": "Promotion",
                "description": "Product promotion model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Summer Sale"
                    },
                    "description": {
                        "type": "string",
                        "example": "Get 20% off on selected items"
                    },
                    "discount_type": {
                        "type": "string",
                        "enum": [
                            "percentage",
                            "fixed"
                        ],
                        "example": "percentage"
                    },
                    "discount_value": {
                        "type": "number",
                        "format": "float",
                        "example": 20
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "badge_text": {
                        "type": "string",
                        "example": "20% OFF"
                    },
                    "badge_color": {
                        "type": "string",
                        "example": "#ff0000"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "first_name": {
                        "type": "string",
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": false
                    },
                    "google_id": {
                        "type": "string",
                        "example": "123456789",
                        "nullable": true
                    },
                    "avatar": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg",
                        "nullable": true
                    },
                    "privacy_policy_accepted": {
                        "type": "boolean",
                        "example": true
                    },
                    "privacy_policy_accepted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "terms_of_service_accepted": {
                        "type": "boolean",
                        "example": true
                    },
                    "terms_of_service_accepted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "role": {
                        "type": "string",
                        "example": "user"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "display_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "is_google_user": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}